Moved StudentOpportunitiesSearchPageComponent to components folder, fixed preliminary unit tests
