+++ client/src/app/components/student-opportunites-search-page/student-opportunites-search-page.component.css
