+++ client/src/app/components/student-opportunites-search-page/student-opportunites-search-page.component.spec.ts
@@ -1,6 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { StudentOpportunitesSearchPageComponent } from './student-opportunites-search-page.component';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { FormsModule } from '@angular/forms';
 
 describe('StudentOpportunitesSearchPageComponent', () => {
   let component: StudentOpportunitesSearchPageComponent;
@@ -8,7 +10,8 @@ describe('StudentOpportunitesSearchPageComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [StudentOpportunitesSearchPageComponent]
+      declarations: [StudentOpportunitesSearchPageComponent],
+      imports: [HttpClientTestingModule, FormsModule],
     });
     fixture = TestBed.createComponent(StudentOpportunitesSearchPageComponent);
     component = fixture.componentInstance;
