+++ client/src/app/controllers/student-dashboard-controller/student-dashboard.service.spec.ts
@@ -1,12 +1,15 @@
 import { TestBed } from '@angular/core/testing';
 
 import { StudentDashboardService } from './student-dashboard.service';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('StudentDashboardService', () => {
   let service: StudentDashboardService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [HttpClientTestingModule],
+    });
     service = TestBed.inject(StudentDashboardService);
   });
